requirements.txt
torch
torchvision
pandas
scikit-learn
tensorboard
pyyaml
tqdm # Optional, but used for progress bars in trainer/tester